Rmsd defined in updater but code is broken
