+++ src/main/java/core/Vertex.java
@@ -15,6 +15,8 @@ public class Vertex extends Sphere {
 
     private Point3D velocity = new Point3D(0,0,0);
 
+    private Object luggage;
+
     public Point3D getPoint3D(){
         return new Point3D(super.getTranslateX(),super.getTranslateY(),super.getTranslateZ());
     }
@@ -71,6 +73,14 @@ public class Vertex extends Sphere {
 
     public void setVelocity(Point3D velocity) {
         this.velocity = velocity;
+    }
+
+    public Object getLuggage() {
+        return luggage;
+    }
+
+    public void setLuggage(Object luggage) {
+        this.luggage = luggage;
     }
 
 //    @Override
